requirements.txt
arabic-reshaper==3.0.0
asgiref==3.5.2
asn1crypto==1.5.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.1.2
html5lib==1.1
idna==3.4
lxml==4.9.3
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.2.0
psycopg2==2.9.7
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.6.0
pypdf==3.16.2
pypng==0.20220715.0
python-bidi==0.4.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.5
PyYAML==6.0.1
qrcode==7.4.2
reportlab==3.6.13
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uritools==4.0.2
urllib3==1.26.12
webencodings==0.5.1
xhtml2pdf==0.2.11